moved data into proper dir
